+++ encore_api_cli/cli.py
@@ -4,6 +4,7 @@ from encore_api_cli import __version__
 from encore_api_cli.commands.analysis import cli as analysis
 from encore_api_cli.commands.analyze import cli as analyze
 from encore_api_cli.commands.configure import cli as configure
+from encore_api_cli.commands.download import cli as download
 from encore_api_cli.commands.draw import cli as draw
 from encore_api_cli.commands.image import cli as image
 from encore_api_cli.commands.keypoint import cli as keypoint
@@ -17,6 +18,7 @@ from encore_api_cli.commands.upload import cli as upload
         analysis,
         analyze,
         configure,
+        download,
         draw,
         image,
         keypoint,